+++ src/redux/reducers/instructionsReducers.js
@@ -1,5 +1,5 @@
-export const setInstructions = (_state, action) => {
-  return action.payload;
+export const setInstructions = (state, action) => {
+  return { ...state, ...action.payload };
 };
 
 export const setCurrentStepPosition = (state, action) => {